Refactoring + Enhance UTs + Enhance doc
